Add SVG and PNG favicons, update HTML template links
This commit introduces a new SVG favicon and a PNG favicon to the application. The HTML template has been updated to link to these new favicon files, enhancing the visual branding of the application. The previous favicon link has been modified to support both SVG and ICO formats for better compatibility.